Port or comment out everything so that it compiles. UI files are not ported yet.

+++ src/core/gui/PdfFloatingToolbox.h
@@ -36,7 +36,7 @@ public:
 public:
     /// Show the toolbox at the provided coordinates (in widget-coordinates).
     /// Must have an active selection.
-    void show(int x, int y);
+    void show(double x, double y);
 
     /// Hide the floating toolbox widget (keeping the current selection).
     void hide();
@@ -90,7 +90,7 @@ private:
     std::unique_ptr<PdfElemSelection> pdfElemSelection;
 
     struct {
-        int x;
-        int y;
+        double x;
+        double y;
     } position;
 };
